+++ srcs/settings/settings.hpp
@@ -10,7 +10,7 @@ namespace ws {
 
 struct Settings {
     std::vector<ServerSettings> servers;
-    ServerSettings resolve_settings_hostname(const HttpRequest &request, int port);
+    ServerSettings resolve_settings_hostname(const HttpRequest &request, size_t port);
 };
 
 bool operator==(const Settings &lhs, const Settings &rhs);
